+++ HISTORY.md
@@ -34,6 +34,7 @@
 ### 2022-04-04
 * objects: added speech-balloon-value-association support for ASK menu items
 * extensions: added new extension primitive "loc_translate(text)"
+* extensions: added more extension translation primitives, under construction
 
 ### 2022-04-03
 * objects: added shortcut support for ASK menu items